moved demo stuff to demo package
